requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
rope==0.17.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
